🗃Refactor: EVENT 타입 정의
