+++ src/common/utils/utils.ts
@@ -1,4 +1,5 @@
 import { CustomEventListener } from '@/@types';
+import { EventName } from '@/constant';
 
 export function getRandomRange(start: number, end: number): number {
   return Math.floor(Math.random() * (end - start + 1) + start);
@@ -16,11 +17,11 @@ export function createCustomElement(name: string, customElementConstructor: Cust
   }
 }
 
-export function dispatchCustomEvent(eventName: string, detail: any = {}): void {
+export function dispatchCustomEvent(eventName: EventName, detail: any = {}): void {
   dispatchEvent(new CustomEvent(eventName, { detail }));
 }
 
-export function addCustomEventListener(eventName: string, callback: CustomEventListener): void {
+export function addCustomEventListener(eventName: EventName, callback: CustomEventListener): void {
   addEventListener(eventName, e => {
     callback(e as CustomEvent);
   });
